pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-clusterops"
version = "0.1.0"
description = "ClusterOps: Thermal GPU Balancer environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",
]
training = [
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "stable-baselines3>=2.0.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]